Robustify get_executable_path on macosx. Move defer stuff to general.h
