+++ src/microsoft_craziness.h
@@ -120,26 +120,26 @@ void free_resources(Find_Result *result) {
 
 // Defer macro/thing.
 
-#define CONCAT_INTERNAL(x,y) x##y
-#define CONCAT(x,y) CONCAT_INTERNAL(x,y)
+// #define CONCAT_INTERNAL(x,y) x##y
+// #define CONCAT(x,y) CONCAT_INTERNAL(x,y)
 
-template<typename T>
-struct ExitScope {
-    T lambda;
-    ExitScope(T lambda):lambda(lambda){}
-    ~ExitScope(){lambda();}
-    ExitScope(const ExitScope&);
-  private:
-    ExitScope& operator =(const ExitScope&);
-};
+// template<typename T>
+// struct ExitScope {
+//     T lambda;
+//     ExitScope(T lambda):lambda(lambda){}
+//     ~ExitScope(){lambda();}
+//     ExitScope(const ExitScope&);
+//   private:
+//     ExitScope& operator =(const ExitScope&);
+// };
  
-class ExitScopeHelp {
-  public:
-    template<typename T>
-        ExitScope<T> operator+(T t){ return t;}
-};
+// class ExitScopeHelp {
+//   public:
+//     template<typename T>
+//         ExitScope<T> operator+(T t){ return t;}
+// };
  
-#define defer const auto& CONCAT(defer__, __LINE__) = ExitScopeHelp() + [&]()
+// #define defer const auto& CONCAT(defer__, __LINE__) = ExitScopeHelp() + [&]()
 
 
 // COM objects for the ridiculous Microsoft craziness.
